+++ src/cq/codefetcher.h
@@ -16,10 +16,10 @@ namespace lh{
     {
         private:
             string base_filename;
-            uint32_t M, K, num_docs = 0, key_bytes, offset_bytes, token_bytes;
-            uint32_t number_of_files;
+            size_t M, K, num_docs = 0, key_bytes, offset_bytes, token_bytes;
+            size_t number_of_files;
             vector<ifstream *> file_ptrs;
-            unordered_map<string, int> key_offset_store;
+            unordered_map<string, size_t> key_offset_store;
 
             void load_metadata();
             void initialize_file_ptrs();
@@ -30,7 +30,7 @@ namespace lh{
 
         public:
             // Constructor
-            CodeFetcher(string filename, uint32_t number_of_files);
+            CodeFetcher(string filename, size_t number_of_files);
             ~CodeFetcher();
             unordered_map<string, vector<vector<std::size_t>>> get_codes(vector<string> document_ids);
     };
